+++ Runtime/Views/Effects.cs
@@ -1,7 +1,9 @@
 using System;
 using System.Collections;
 using UnityEngine;
+#if USE_TMP
 using TMPro;
+#endif
 
 namespace Yarn.Unity
 {
@@ -126,6 +128,7 @@ namespace Yarn.Unity
             stopToken?.Complete();
         }
 
+#if USE_TMP
         /// <summary>
         /// A coroutine that gradually reveals the text in a <see
         /// cref="TextMeshProUGUI"/> object over time.
@@ -205,5 +208,6 @@ namespace Yarn.Unity
 
             stopToken?.Complete();
         }
+#endif
     }
 }