+++ Runtime/Views/LineView.cs
@@ -1,3 +1,4 @@
+#if USE_TMP
 using System;
 using System.Collections;
 using UnityEngine;
@@ -471,3 +472,12 @@ namespace Yarn.Unity
         }
     }
 }
+#else
+namespace Yarn.Unity
+{
+    /// <summary>
+    /// Required to draw an inspector saying that TextMeshPro must be added for this component to be supported.
+    /// </summary>
+    public class LineView : DialogueViewBase { }
+}
+#endif